axis_stream_to_pkt_backpressured: fix enable and fifos_reset bug
This fixes a bug that was introduced with the addition of the fifos_reset
signal, which causes the internal FIFOs to be reset so that stale
samples are dropped when the module is disabled.

The bug is that when enable is asserted, fifos_reset takes 2 clock
cycles to deassert, because first output_state must change to
S_OUTPUT_HEADER, and then the fifos_reset flip-flop must update to 0.
If samples are available at the input every clock cycle, then the
first clock cycle the input state machine is in S_IDLE and accepts
one input sample, passing to S_INPUT_PHASE_2. The second clock cycle,
a new input sample is accepted and the two samples are attempted to
be written to the sample_fifo. However this FIFO is still in reset,
so it discards the write. This causes a mismatch between the input_count
(since both samples have been counted) and the number of samples in
sample_fifo.

To fix this bug && !enable is added to the input of the fifos_reset,
flip-flop. This causes fifos_reset to go low one cycle earlier, so
no data is lost.